+++ src/main/java/com/google/api/gax/core/AsyncPage.java
@@ -29,6 +29,8 @@
  */
 package com.google.api.gax.core;
 
+import com.google.api.common.ApiFuture;
+
 /**
  * An AsyncPage extends a {@link Page} object with a getNextPageAsync method.
  */
