+++ src/main/java/com/google/api/gax/grpc/AbstractPage.java
@@ -29,9 +29,9 @@
  */
 package com.google.api.gax.grpc;
 
-import com.google.api.gax.core.ApiFunction;
-import com.google.api.gax.core.ApiFuture;
-import com.google.api.gax.core.ApiFutures;
+import com.google.api.common.ApiFunction;
+import com.google.api.common.ApiFuture;
+import com.google.api.common.ApiFutures;
 import com.google.api.gax.core.AsyncPage;
 import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.Iterables;
