+++ src/main/java/com/google/api/gax/grpc/BatchExecutor.java
@@ -29,11 +29,11 @@
  */
 package com.google.api.gax.grpc;
 
+import com.google.api.common.ApiFuture;
+import com.google.api.common.ApiFutureCallback;
+import com.google.api.common.ApiFutures;
 import com.google.api.gax.batching.PartitionKey;
 import com.google.api.gax.batching.ThresholdBatchReceiver;
-import com.google.api.gax.core.ApiFuture;
-import com.google.api.gax.core.ApiFutureCallback;
-import com.google.api.gax.core.ApiFutures;
 import com.google.common.base.Preconditions;
 import java.util.List;
