+++ src/main/java/com/google/api/gax/grpc/ExceptionTransformingCallable.java
@@ -29,10 +29,10 @@
  */
 package com.google.api.gax.grpc;
 
-import com.google.api.gax.core.AbstractApiFuture;
-import com.google.api.gax.core.ApiFuture;
-import com.google.api.gax.core.ApiFutureCallback;
-import com.google.api.gax.core.ApiFutures;
+import com.google.api.common.AbstractApiFuture;
+import com.google.api.common.ApiFuture;
+import com.google.api.common.ApiFutureCallback;
+import com.google.api.common.ApiFutures;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 import io.grpc.Status;
