+++ src/main/java/com/google/api/gax/protobuf/ProtoReflectionUtil.java
@@ -29,6 +29,7 @@
  */
 package com.google.api.gax.protobuf;
 
+import com.google.api.pathtemplate.ValidationException;
 import com.google.common.base.CaseFormat;
 import com.google.common.base.Throwables;
 import com.google.common.io.Files;
