+++ src/main/java/com/google/api/gax/retrying/ExponentialRetryAlgorithm.java
@@ -31,7 +31,7 @@ package com.google.api.gax.retrying;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.google.api.gax.core.ApiClock;
+import com.google.api.common.ApiClock;
 import com.google.api.gax.core.RetrySettings;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
