+++ src/main/java/com/google/api/gax/retrying/RetryingFuture.java
@@ -29,7 +29,7 @@
  */
 package com.google.api.gax.retrying;
 
-import com.google.api.gax.core.ApiFuture;
+import com.google.api.common.ApiFuture;
 import java.util.concurrent.Callable;
 
 /**
