+++ src/main/java/com/google/longrunning/PagedResponseWrappers.java
@@ -29,9 +29,9 @@
  */
 package com.google.longrunning;
 
-import com.google.api.gax.core.ApiFunction;
-import com.google.api.gax.core.ApiFuture;
-import com.google.api.gax.core.ApiFutures;
+import com.google.api.common.ApiFunction;
+import com.google.api.common.ApiFuture;
+import com.google.api.common.ApiFutures;
 import com.google.api.gax.grpc.AbstractFixedSizeCollection;
 import com.google.api.gax.grpc.AbstractPage;
 import com.google.api.gax.grpc.AbstractPagedListResponse;
