+++ src/test/java/com/google/api/gax/grpc/OperationFutureTest.java
@@ -29,7 +29,7 @@
  */
 package com.google.api.gax.grpc;
 
-import com.google.api.gax.core.SettableApiFuture;
+import com.google.api.common.SettableApiFuture;
 import com.google.api.gax.testing.MockGrpcService;
 import com.google.api.gax.testing.MockServiceHelper;
 import com.google.common.truth.Truth;
