+++ src/test/java/com/google/api/gax/grpc/UnaryCallableTest.java
@@ -29,20 +29,20 @@
  */
 package com.google.api.gax.grpc;
 
+import com.google.api.common.ApiFunction;
+import com.google.api.common.ApiFuture;
+import com.google.api.common.ApiFutures;
+import com.google.api.common.FakeApiClock;
 import com.google.api.gax.batching.BatchingSettings;
 import com.google.api.gax.batching.PartitionKey;
 import com.google.api.gax.batching.RequestBuilder;
-import com.google.api.gax.core.ApiFunction;
-import com.google.api.gax.core.ApiFuture;
-import com.google.api.gax.core.ApiFutures;
-import com.google.api.gax.core.FakeApiClock;
 import com.google.api.gax.core.FixedSizeCollection;
 import com.google.api.gax.core.FlowControlSettings;
 import com.google.api.gax.core.FlowController.LimitExceededBehavior;
 import com.google.api.gax.core.Page;
 import com.google.api.gax.core.RetrySettings;
 import com.google.api.gax.core.TrackedFlowController;
-import com.google.api.gax.protobuf.ValidationException;
+import com.google.api.pathtemplate.ValidationException;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.truth.Truth;
